query normalization and score done

+++ src/main/java/SearchEngine/QueryProcessing/Query.java
@@ -9,7 +9,7 @@ import java.util.Map;
 
 public class Query {
     private int query_id;
-    private Map<Integer, Integer> doc_scores;
+    private Map<Integer, Double> doc_scores;
 
     /**
      * Creates a new instance of a query given an integer value that represents the query id
@@ -33,7 +33,7 @@ public class Query {
      * Return a map containing pair values (doc_id, doc_score) for this query
      * @return query's doc scores
      */
-    public Map<Integer, Integer> getDoc_scores() {
+    public Map<Integer, Double> getDoc_scores() {
         return doc_scores;
     }
 
@@ -42,7 +42,7 @@ public class Query {
      * @param docID the document id
      * @param number the value to be added to the score
      */
-    public void increaseDocScore(int docID, int number) {
+    public void increaseDocScore(int docID, double number) {
         doc_scores.put(docID, doc_scores.containsKey(docID) ? doc_scores.get(docID) + number : number);
     }
